Add script to remove queries for fixed bugs
This automates the process to remove old queries
for fixed bugs. It's a bit conservative to start
so it doesn't check for open reviews nor does it
filter out affected projects with non-Fix* status
on the bug. It can be made more robust once we're
confident in how it works and play with it on the
open queries.

Change-Id: Iaaf17892804453b99a846be27457c88e5a8f8a55