Handle ElasticHttpError in graph generation
Getting elasticsearch data for bug 1708704 is failing
in the check queue with:

  pyelasticsearch.exceptions.ElasticHttpError: \
  (500, 'ArrayIndexOutOfBoundsException[null]')

This might have to do with the size of the resulting
messages from the hits on the tripleo and kolla jobs,
I'm not sure.

What's clear though is the graph generation is blowing
up in the check queue on that bug but not the gate queue,
maybe due to a smaller result set, so this adds some
error handling in the graph generation for when a specific
bug query fails so it does not halt the entire build of the
graph.

Change-Id: Ibe18c9cccc421a6549a18148f1a2ce3c1e4339d4