Simplify 1708704's query
The previous query for 1708704 was causing e-r to OOM and resulting in a
lack of graph data. This is beacuse tripleo's logstash.txt file is not
getting parsed properly and ends up with the entire file as the single
event. Then e-r downloads all those file copies and fills its memory and
breaks. We can work around this by only looking in job-output.txt files
for this bug. Then when tripleo's fix has flushed the bad events out (10
days after fix merges) we can revert this change.

Change-Id: Id619f90ffe84b3d4de334ea4b17026b9b3239d33